refactor: extract autocomplete to a separate package
This also entails extracting the search as a dedicated component, since it's currently not behaving like the input in the fields
